feat: Migrate VSCode extensions and Cursor settings to mixin architecture
**VSCode Extensions Migration:**
- 58% code reduction: 359 → 152 lines
- Migrated to ExtensionsPlugin<VSCodeExtensionsCore>
- Automatic command execution with version flags
- Comprehensive restore functionality with alternative file detection

**Cursor Settings Migration:**
- 46% code reduction: 388 → 211 lines
- Migrated to SettingsPlugin<CursorCore>
- Platform-specific directory resolution (macOS/Windows/Linux)
- Full JSON/JSONC validation and restore support

**Architecture Improvements:**
- Fixed sync/async API boundary in SettingsCore trait
- get_settings_dir() now sync for lightweight path resolution
- read_settings() remains async for file I/O operations
- Improved test compatibility with Tokio runtime contexts

**Total Progress:**
- 3 plugins migrated: VSCode settings (53% reduction), VSCode extensions (58% reduction), Cursor settings (46% reduction)
- Average code reduction: 52%
- All functionality preserved with enhanced consistency
- Zero breaking changes to plugin interfaces
