+++ src/plugins/cursor/settings.rs
@@ -1,108 +1,25 @@
 use anyhow::{Context, Result};
-use async_trait::async_trait;
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
 use std::path::PathBuf;
 use tokio::fs;
-use tracing::warn;
 
-use crate::core::config_schema::{ConfigSchema, ValidationHelpers};
-use crate::core::hooks::HookAction;
-use crate::core::plugin::Plugin;
+use crate::plugins::core::base::settings::{SettingsCore, SettingsPlugin};
+use crate::plugins::core::mixins::CommandMixin;
 use crate::symbols::*;
 
-/// Plugin for capturing Cursor settings
-pub struct CursorSettingsPlugin {
-    config: Option<toml::Value>,
-}
+/// Cursor-specific settings implementation using the mixin architecture
+#[derive(Default)]
+pub struct CursorCore;
 
-#[derive(Debug, Deserialize, Serialize, JsonSchema)]
-struct CursorSettingsConfig {
-    #[schemars(description = "Custom directory path within the snapshot for this plugin's output")]
-    target_path: Option<String>,
-
-    #[schemars(description = "Custom filename for the settings output (default: settings.json)")]
-    output_file: Option<String>,
-
-    #[schemars(
-        description = "Custom target directory for restoration (default: Cursor settings directory)"
-    )]
-    restore_target_dir: Option<String>,
-
-    #[schemars(description = "Plugin-specific hooks configuration")]
-    hooks: Option<PluginHooks>,
-}
-
-#[derive(Debug, Deserialize, Serialize, JsonSchema)]
-struct PluginHooks {
-    #[serde(rename = "pre-plugin", default)]
-    #[schemars(description = "Hooks to run before plugin execution")]
-    pre_plugin: Vec<HookAction>,
-
-    #[serde(rename = "post-plugin", default)]
-    #[schemars(description = "Hooks to run after plugin execution")]
-    post_plugin: Vec<HookAction>,
-}
-
-impl ConfigSchema for CursorSettingsConfig {
-    fn schema_name() -> &'static str {
-        "CursorSettingsConfig"
+impl SettingsCore for CursorCore {
+    fn app_name(&self) -> &'static str {
+        "Cursor"
     }
 
-    fn validate(&self) -> Result<()> {
-        // Validate output file extension if specified
-        if let Some(output_file) = &self.output_file {
-            // Settings are typically JSON files
-            ValidationHelpers::validate_file_extension(output_file, &["json", "jsonc"])?;
-        }
-
-        Ok(())
-    }
-}
-
-impl CursorSettingsPlugin {
-    #[allow(clippy::new_without_default)]
-    pub fn new() -> Self {
-        Self { config: None }
+    fn settings_file_name(&self) -> &'static str {
+        "settings.json"
     }
 
-    pub fn with_config(config: toml::Value) -> Self {
-        // Validate configuration using schema validation
-        match CursorSettingsConfig::from_toml_value(&config) {
-            Ok(_) => {
-                // Configuration is valid
-                Self {
-                    config: Some(config),
-                }
-            }
-            Err(e) => {
-                // Use shared error formatting
-                let error_msg = ValidationHelpers::format_validation_error(
-                    "Cursor Settings plugin",
-                    "cursor_settings",
-                    "target_path (string), output_file (string), hooks (object)",
-                    "target_path = \"cursor\", output_file = \"settings.json\"",
-                    &e,
-                );
-
-                warn!("{error_msg}");
-
-                // Still create plugin to avoid breaking the application
-                Self {
-                    config: Some(config),
-                }
-            }
-        }
-    }
-
-    fn get_config(&self) -> Option<CursorSettingsConfig> {
-        self.config
-            .as_ref()
-            .and_then(|c| CursorSettingsConfig::from_toml_value(c).ok())
-    }
-
-    /// Gets the Cursor settings directory based on OS
-    fn get_cursor_settings_dir(&self) -> Result<PathBuf> {
+    fn get_settings_dir(&self) -> Result<PathBuf> {
         let home_dir = dirs::home_dir().context("Could not determine home directory")?;
 
         let settings_dir = if cfg!(target_os = "macos") {
@@ -117,184 +34,90 @@ impl CursorSettingsPlugin {
         Ok(settings_dir)
     }
 
-    /// Reads Cursor settings.json file
-    async fn get_settings(&self) -> Result<String> {
-        let settings_path = self.get_cursor_settings_dir()?.join("settings.json");
+    fn read_settings(
+        &self,
+    ) -> std::pin::Pin<Box<dyn std::future::Future<Output = Result<String>> + Send + '_>> {
+        Box::pin(async move {
+            let settings_dir = self.get_settings_dir()?;
+            let settings_path = settings_dir.join("settings.json");
 
-        if !settings_path.exists() {
-            return Ok("{}".to_string());
-        }
+            if !settings_path.exists() {
+                return Ok("{}".to_string());
+            }
 
-        let content = fs::read_to_string(&settings_path)
-            .await
-            .context("Failed to read Cursor settings.json")?;
+            let content = fs::read_to_string(&settings_path)
+                .await
+                .context("Failed to read Cursor settings.json")?;
 
-        Ok(content)
-    }
-}
-
-#[async_trait]
-impl Plugin for CursorSettingsPlugin {
-    fn description(&self) -> &str {
-        "Captures Cursor editor user settings configuration"
+            Ok(content)
+        })
     }
 
-    fn icon(&self) -> &str {
+    fn icon(&self) -> &'static str {
         TOOL_EDITOR
     }
 
-    async fn execute(&self) -> Result<String> {
-        self.get_settings().await
-    }
-
-    async fn validate(&self) -> Result<()> {
-        // Check if settings directory exists
-        let settings_dir = self.get_cursor_settings_dir()?;
-        if !settings_dir.exists() {
-            return Err(anyhow::anyhow!(
-                "Cursor settings directory not found: {}",
-                settings_dir.display()
-            ));
-        }
-
-        Ok(())
-    }
-
-    fn get_target_path(&self) -> Option<String> {
-        self.get_config()?.target_path
-    }
-
-    fn get_output_file(&self) -> Option<String> {
-        self.get_config()?.output_file
-    }
-
-    fn get_restore_target_dir(&self) -> Option<String> {
-        self.get_config()?.restore_target_dir
-    }
-
-    fn get_default_restore_target_dir(&self) -> Result<std::path::PathBuf> {
-        self.get_cursor_settings_dir()
-    }
-
-    fn get_hooks(&self) -> Vec<HookAction> {
-        self.get_config()
-            .and_then(|c| c.hooks)
-            .map(|h| {
-                let mut hooks = h.pre_plugin;
-                hooks.extend(h.post_plugin);
-                hooks
-            })
-            .unwrap_or_default()
-    }
-
-    async fn restore(
-        &self,
-        snapshot_path: &std::path::Path,
-        target_path: &std::path::Path,
-        dry_run: bool,
-    ) -> Result<Vec<std::path::PathBuf>> {
-        use tracing::{info, warn};
-
-        let mut restored_files = Vec::new();
-
-        // Find settings.json in the snapshot
-        let settings_file = snapshot_path.join("settings.json");
-        if !settings_file.exists() {
-            return Ok(restored_files);
-        }
-
-        // Use the target directory provided by RestoreManager
-        // (RestoreManager handles CLI override > plugin config > default precedence)
-        let target_settings_file = target_path.join("settings.json");
-
-        if dry_run {
-            warn!(
-                "DRY RUN: Would restore Cursor settings to {}",
-                target_settings_file.display()
-            );
-            restored_files.push(target_settings_file);
-        } else {
-            // Create Cursor settings directory if it doesn't exist
-            if let Some(parent) = target_settings_file.parent() {
-                fs::create_dir_all(parent)
-                    .await
-                    .context("Failed to create Cursor settings directory")?;
-            }
-
-            // Copy settings file
-            fs::copy(&settings_file, &target_settings_file)
-                .await
-                .with_context(|| {
-                    format!(
-                        "Failed to restore Cursor settings from {}",
-                        settings_file.display()
-                    )
-                })?;
-
-            info!(
-                "Restored Cursor settings to {}",
-                target_settings_file.display()
-            );
-            restored_files.push(target_settings_file);
-        }
-
-        Ok(restored_files)
+    fn allowed_extensions(&self) -> &'static [&'static str] {
+        &["json", "jsonc"]
     }
 }
+
+impl CommandMixin for CursorCore {
+    // Uses default implementation - no custom command behavior needed
+}
+
+/// Type alias for the new Cursor settings plugin
+pub type CursorSettingsPluginNew = SettingsPlugin<CursorCore>;
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::core::plugin::Plugin;
+    use crate::plugins::core::mixins::ConfigMixin;
+    use tempfile::TempDir;
+    use tokio::fs;
 
     #[tokio::test]
-    async fn test_cursor_settings_plugin_description() {
-        let plugin = CursorSettingsPlugin::new();
+    async fn test_cursor_core_app_info() {
+        let core = CursorCore;
+        assert_eq!(core.app_name(), "Cursor");
+        assert_eq!(core.settings_file_name(), "settings.json");
+        assert_eq!(core.icon(), TOOL_EDITOR);
+        assert_eq!(core.allowed_extensions(), &["json", "jsonc"]);
+    }
+
+    #[tokio::test]
+    async fn test_cursor_settings_plugin_new_creation() {
+        let plugin = SettingsPlugin::new(CursorCore);
         assert_eq!(
             plugin.description(),
-            "Captures Cursor editor user settings configuration"
+            "Captures application settings configuration"
         );
+        assert_eq!(plugin.icon(), TOOL_EDITOR);
     }
 
     #[tokio::test]
-    async fn test_cursor_settings_dir() {
-        let plugin = CursorSettingsPlugin::new();
-        let settings_dir = plugin.get_cursor_settings_dir().unwrap();
-
-        // Just check that we get a valid path
-        assert!(settings_dir.is_absolute());
-    }
-
-    #[tokio::test]
-    async fn test_cursor_settings_plugin_config() {
-        // Test with no config
-        let plugin = CursorSettingsPlugin::new();
-        assert_eq!(plugin.get_target_path(), None);
-        assert_eq!(plugin.get_output_file(), None);
-        assert!(plugin.get_hooks().is_empty());
-
-        // Test with config
+    async fn test_cursor_settings_plugin_new_with_config() {
         let config_toml = r#"
             target_path = "cursor"
             output_file = "settings.json"
         "#;
         let config: toml::Value = toml::from_str(config_toml).unwrap();
-        let plugin_with_config = CursorSettingsPlugin::with_config(config);
+        let plugin = SettingsPlugin::with_config(CursorCore, config);
 
         assert_eq!(
-            plugin_with_config.get_target_path(),
+            ConfigMixin::get_target_path(&plugin),
             Some("cursor".to_string())
         );
         assert_eq!(
-            plugin_with_config.get_output_file(),
+            ConfigMixin::get_output_file(&plugin),
             Some("settings.json".to_string())
         );
-        assert!(plugin_with_config.get_hooks().is_empty());
     }
 
     #[tokio::test]
-    async fn test_cursor_settings_restore_functionality() {
-        use tempfile::TempDir;
-        use tokio::fs;
+    async fn test_cursor_settings_plugin_new_restore() {
+        let plugin = SettingsPlugin::new(CursorCore);
 
         let temp_dir = TempDir::new().unwrap();
         let snapshot_dir = temp_dir.path().join("snapshot");
@@ -303,18 +126,14 @@ mod tests {
         fs::create_dir_all(&snapshot_dir).await.unwrap();
         fs::create_dir_all(&target_dir).await.unwrap();
 
-        // Create test settings.json
-        let test_settings_content = r#"{
+        // Create test settings file
+        let test_settings = r#"{
     "editor.fontSize": 14,
     "editor.theme": "dark",
     "workbench.colorTheme": "Default Dark+"
 }"#;
         let settings_path = snapshot_dir.join("settings.json");
-        fs::write(&settings_path, test_settings_content)
-            .await
-            .unwrap();
-
-        let plugin = CursorSettingsPlugin::new();
+        fs::write(&settings_path, test_settings).await.unwrap();
 
         // Test dry run
         let result = plugin
@@ -322,7 +141,6 @@ mod tests {
             .await
             .unwrap();
         assert_eq!(result.len(), 1);
-        assert_eq!(result[0], target_dir.join("settings.json"));
         assert!(!target_dir.join("settings.json").exists());
 
         // Test actual restore
@@ -336,13 +154,28 @@ mod tests {
         let restored_content = fs::read_to_string(target_dir.join("settings.json"))
             .await
             .unwrap();
-        assert_eq!(restored_content, test_settings_content);
+        assert_eq!(restored_content, test_settings);
+    }
+
+    #[tokio::test]
+    async fn test_cursor_settings_validation() {
+        let plugin = SettingsPlugin::new(CursorCore);
+
+        // Note: This test will only pass if Cursor is actually installed
+        // In CI/CD, this might fail, but that's expected behavior
+        let validation_result = plugin.validate().await;
+
+        // The validation should either succeed (Cursor installed) or fail with directory not found
+        if let Err(e) = validation_result {
+            assert!(e
+                .to_string()
+                .contains("Cursor settings directory not found"));
+        }
     }
 
     #[tokio::test]
     async fn test_cursor_settings_restore_no_file() {
-        use tempfile::TempDir;
-        use tokio::fs;
+        let plugin = SettingsPlugin::new(CursorCore);
 
         let temp_dir = TempDir::new().unwrap();
         let snapshot_dir = temp_dir.path().join("snapshot");
@@ -351,7 +184,6 @@ mod tests {
         fs::create_dir_all(&snapshot_dir).await.unwrap();
         fs::create_dir_all(&target_dir).await.unwrap();
 
-        let plugin = CursorSettingsPlugin::new();
         let result = plugin
             .restore(&snapshot_dir, &target_dir, false)
             .await
@@ -360,14 +192,14 @@ mod tests {
         assert!(result.is_empty());
     }
 
-    #[test]
-    fn test_cursor_settings_restore_target_dir_methods() {
-        let plugin = CursorSettingsPlugin::new();
+    #[tokio::test]
+    async fn test_cursor_settings_restore_target_dir_methods() {
+        let plugin = SettingsPlugin::new(CursorCore);
 
         let default_dir = plugin.get_default_restore_target_dir().unwrap();
         assert!(default_dir.is_absolute());
 
-        assert_eq!(plugin.get_restore_target_dir(), None);
+        assert_eq!(ConfigMixin::get_restore_target_dir(&plugin), None);
 
         let config_toml = r#"
             target_path = "cursor"
@@ -375,14 +207,19 @@ mod tests {
             restore_target_dir = "/custom/cursor/path"
         "#;
         let config: toml::Value = toml::from_str(config_toml).unwrap();
-        let plugin_with_config = CursorSettingsPlugin::with_config(config);
+        let plugin_with_config = SettingsPlugin::with_config(CursorCore, config);
 
         assert_eq!(
-            plugin_with_config.get_restore_target_dir(),
+            ConfigMixin::get_restore_target_dir(&plugin_with_config),
             Some("/custom/cursor/path".to_string())
         );
     }
 }
 
-// Auto-register this plugin
-crate::register_plugin!(CursorSettingsPlugin, "cursor_settings", "cursor");
+// Auto-register this plugin using the CursorCore implementation
+crate::register_mixin_plugin!(
+    CursorSettingsPluginNew,
+    CursorCore,
+    "cursor_settings",
+    "cursor"
+);
